Update database functionality
A couple of accounts were updated, but they were minor changes that don't need to be noted quite yet. (However, I should've done all of this in smaller commits, my "do everything at once" approach is starting to show its flaws).

Added a brief explanation for the database.

Updated the database helper functions to no longer requiring a class input. Created an account for restoring customer accounts to their internal arrayLists when the database is being restored. Added a function for getting a specific account from a specific arrayList.